restyled everything

+++ asynctnt/iproto/bit.pxd
@@ -1,5 +1,6 @@
-from libc.stdint cimport uint64_t, uint32_t, uint16_t
+from libc.stdint cimport uint16_t, uint32_t, uint64_t
 from libc.string cimport memcpy
+
 
 cdef inline uint64_t load_u64(const void * p):
     cdef:
